+++ python/tools/drf_sti.py
@@ -154,7 +154,7 @@ class DataPlotter(object):
             sti_psd_data = np.zeros([self.control.num_fft, self.control.bins], np.float)
             sti_times = np.zeros([self.control.bins], np.complex128)
 
-            for b in np.arange(self.control.bins):
+            for b in np.arange(self.control.bins, dtype=np.int_):
 
                 if self.control.verbose:
                     print(
@@ -237,7 +237,10 @@ class DataPlotter(object):
             # plot dates
 
             tick_spacing = np.arange(
-                self.control.bins / 8, self.control.bins, self.control.bins / 8
+                self.control.bins / 8,
+                self.control.bins,
+                self.control.bins / 8,
+                dtype=np.int_,
             )
             ax.set_xticks(tick_spacing)
             tick_labels = []
